# backend/openapi-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/messaging/conversations": {
      "get": {
        "description": "Retrieves a paginated list of conversations for a client account with optional filters",
        "operationId": "MessagingController_listConversations",
        "parameters": [
          {
            "name": "clientAccountId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "open",
                "closed",
                "archived"
              ]
            }
          },
          {
            "name": "hasUnread",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 50,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "default": 0,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of conversations retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversationListResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Client account not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List conversations",
        "tags": [
          "Messaging"
        ]
      }
    },
    "/api/messaging/conversations/{id}": {
      "get": {
        "description": "Retrieves detailed information about a specific conversation",
        "operationId": "MessagingController_getConversation",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Conversation ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Conversation details retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversationResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Conversation not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get conversation details",
        "tags": [
          "Messaging"
        ]
      }
    },
    "/api/messaging/conversations/{id}/messages": {
      "get": {
        "description": "Retrieves a paginated list of messages for a specific conversation",
        "operationId": "MessagingController_listMessages",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Conversation ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 100,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "default": 0,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of messages retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageListResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Conversation not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List messages in conversation",
        "tags": [
          "Messaging"
        ]
      },
      "post": {
        "description": "Sends a text message to a participant in a conversation via Instagram",
        "operationId": "MessagingController_sendMessage",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Conversation ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendMessageDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Message sent successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid message or 24-hour response window has expired"
          },
          "403": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Conversation not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Send a message",
        "tags": [
          "Messaging"
        ]
      }
    },
    "/api/messaging/conversations/{id}/read": {
      "patch": {
        "description": "Marks all unread messages in a conversation as read",
        "operationId": "MessagingController_markAsRead",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Conversation ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Conversation marked as read successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversationResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Conversation not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mark conversation as read",
        "tags": [
          "Messaging"
        ]
      }
    },
...
  },
  "info": {
    "title": "Social Selling API",
    "description": "\n# Social Selling Platform API\n\nComplete REST API for Instagram-based social selling platform.\n\n## Features\n- Instagram OAuth integration\n- Product catalog management\n- Direct message automation\n- Post scheduling and analytics\n- Multi-account support\n\n## Authentication\nAll authenticated endpoints require a Bearer token in the Authorization header.\nObtain tokens via POST /api/auth/login endpoint.\n\n## Rate Limiting\n- Authenticated: 1000 requests/hour\n- Unauthenticated: 100 requests/hour\n\n## Session Management\nSessions are managed via HTTP-only cookies and can be tracked per device.\n  ",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Local Development"
    },
    {
      "url": "https://staging-api.socialselling.willianbvsanches.com",
      "description": "Staging Environment"
    },
    {
      "url": "https://api.socialselling.willianbvsanches.com",
      "description": "Production Environment"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "Authorization",
        "description": "Enter JWT access token",
        "in": "header"
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "ssell_session",
        "description": "Session cookie for authenticated requests"
      }
    },
    "schemas": {
...
      "ConversationResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "clientAccountId": {
            "type": "string"
          },
          "platformConversationId": {
            "type": "string"
          },
          "participantPlatformId": {
            "type": "string"
          },
          "participantUsername": {
            "type": "string"
          },
          "participantProfilePic": {
            "type": "string"
          },
          "lastMessageAt": {
            "format": "date-time",
            "type": "string"
          },
          "unreadCount": {
            "type": "number"
          },
          "status": {
            "type": "string",
            "enum": [
              "open",
              "closed",
              "archived"
            ]
          },
          "metadata": {
            "type": "object"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "clientAccountId",
          "platformConversationId",
          "participantPlatformId",
          "unreadCount",
          "status",
          "metadata",
          "createdAt",
          "updatedAt"
        ]
      },
      "ConversationListResponseDto": {
        "type": "object",
        "properties": {
          "conversations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ConversationResponseDto"
            }
          },
          "total": {
            "type": "number"
          },
          "limit": {
            "type": "number"
          },
          "offset": {
            "type": "number"
          }
        },
        "required": [
          "conversations",
          "total",
          "limit",
          "offset"
        ]
      },
      "MessageResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "conversationId": {
            "type": "string"
          },
          "platformMessageId": {
            "type": "string"
          },
          "senderType": {
            "type": "string",
            "enum": [
              "user",
              "customer"
            ]
          },
          "senderPlatformId": {
            "type": "string"
          },
          "messageType": {
            "type": "string",
            "enum": [
              "text",
              "image",
              "video",
              "audio",
              "story_mention",
              "story_reply"
            ]
          },
          "content": {
            "type": "string"
          },
          "mediaUrl": {
            "type": "string"
          },
          "mediaType": {
            "type": "string"
          },
          "isRead": {
            "type": "boolean"
          },
          "sentAt": {
            "format": "date-time",
            "type": "string"
          },
          "deliveredAt": {
            "format": "date-time",
            "type": "string"
          },
          "readAt": {
            "format": "date-time",
            "type": "string"
          },
          "metadata": {
            "type": "object"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "conversationId",
          "platformMessageId",
          "senderType",
          "messageType",
          "isRead",
          "sentAt",
          "metadata",
          "createdAt"
        ]
      },
      "MessageListResponseDto": {
        "type": "object",
        "properties": {
          "messages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MessageResponseDto"
            }
          },
          "total": {
            "type": "number"
          },
          "limit": {
            "type": "number"
          },
          "offset": {
            "type": "number"
          }
        },
        "required": [
          "messages",
          "total",
          "limit",
          "offset"
        ]
      },
      "SendMessageDto": {
        "type": "object",
        "properties": {
          "text": {
            "type": "string",
            "minLength": 1,
            "maxLength": 1000
          }
        },
        "required": [
          "text"
        ]
      }
    }
  }
}